🚚 consolidate cf-core type definitions
