+++ modules/cf-core/src/ethereum/setup-commitment.ts
@@ -1,10 +1,9 @@
 import { Interface } from "ethers/utils";
 
 import { ConditionalTransactionDelegateTarget } from "../contracts";
-import { AppIdentity, NetworkContext } from "../types";
+import { AppIdentity, MultisigOperation, MultisigTransaction, NetworkContext } from "../types";
 
 import { MultisigCommitment } from "./multisig-commitment";
-import { MultisigOperation, MultisigTransaction } from "./types";
 import { appIdentityToHash } from "./utils/app-identity";
 
 const iface = new Interface(ConditionalTransactionDelegateTarget.abi);
