+++ modules/cf-core/src/machine/protocol-runner.ts
@@ -3,15 +3,12 @@ import uuid from "uuid";
 
 import { StateChannel } from "../models";
 import { getProtocolFromName } from "../protocol";
-import { NetworkContext } from "../types";
-
-import { Opcode, Protocol } from "./enums";
-import { MiddlewareContainer } from "./middleware";
 import {
   Context,
   InstallProtocolParams,
   InstallVirtualAppProtocolParams,
   Middleware,
+  NetworkContext,
   ProposeInstallProtocolParams,
   ProtocolMessage,
   SetupProtocolParams,
@@ -19,8 +16,11 @@ import {
   UninstallProtocolParams,
   UninstallVirtualAppProtocolParams,
   UpdateProtocolParams,
-  WithdrawProtocolParams
-} from "./types";
+  WithdrawProtocolParams,
+} from "../types";
+
+import { Opcode, Protocol } from "./enums";
+import { MiddlewareContainer } from "./middleware";
 
 /**
 Type-level mapping from Protocol to Protocol Param
