+++ modules/cf-core/src/protocol/update.ts
@@ -1,7 +1,7 @@
 import { SetStateCommitment } from "../ethereum";
 import { ProtocolExecutionFlow, xkeyKthAddress } from "../machine";
 import { Opcode, Protocol } from "../machine/enums";
-import { Context, ProtocolMessage, UpdateProtocolParams } from "../machine/types";
+import { Context, ProtocolMessage, UpdateProtocolParams } from "../types";
 
 import { UNASSIGNED_SEQ_NO } from "./utils/signature-forwarder";
 import { assertIsValidSignature } from "./utils/signature-validator";
