+++ modules/cf-core/test/integration/install.spec.ts
@@ -3,8 +3,7 @@ import { BigNumber } from "ethers/utils";
 
 import { Node, NULL_INITIAL_STATE_FOR_PROPOSAL } from "../../src";
 import { CONVENTION_FOR_ETH_TOKEN_ADDRESS } from "../../src/constants";
-import { ProposeInstallProtocolParams } from "../../src/machine/types";
-import { InstallMessage, NODE_EVENTS, ProposeMessage } from "../../src/types";
+import { InstallMessage, ProposeInstallProtocolParams, ProposeMessage } from "../../src/types";
 import { NetworkContextForTestSuite } from "../contracts";
 import { toBeLt } from "../machine/integration/bignumber-jest-matcher";
