+++ modules/cf-core/test/machine/unit/ethereum/setup-commitment.spec.ts
@@ -7,7 +7,7 @@ import {
 } from "ethers/utils";
 
 import { SetupCommitment } from "../../../../src/ethereum";
-import { MultisigTransaction } from "../../../../src/ethereum/types";
+import { MultisigTransaction } from "../../../../src/types";
 import { appIdentityToHash } from "../../../../src/ethereum/utils/app-identity";
 import { StateChannel } from "../../../../src/models";
 import { ConditionalTransactionDelegateTarget } from "../../../contracts";
